+++ src/components/Home/Sponsors.js
@@ -1,79 +1,75 @@
 import React from "react";
-import { Col, Row } from "react-bootstrap";
-
-// Sample sponsors data
-const sponsorsData = {
-    "title": "Sponsors",
-    "sponsors": [
-        {
-            "name": "Vanderbilt Brain Institute",
-            "url": "https://medschool.vanderbilt.edu/brain-institute/",
-            "img": "../../images/sponsors/vbi.png",
-            "dims": {
-                "width": 250,
-                "height": 150
-            }
-        },
-        {
-            "name": "VUCS",
-            "url": "https://www.vanderbilt.edu/",
-            "img": "../../images/sponsors/cs.png",
-            "dims": {
-                "width": 250,
-                "height": 150
-            }
-        },
-        {
-            "name": "VUSE",
-            "url": "https://www.vumc.org/",
-            "img": "../../images/sponsors/vuse.png",
-            "dims": {
-                "width": 250,
-                "height": 150
-            }
-        }
-    ]
-};
+import { Container, Row, Col } from "react-bootstrap";
 
 function Sponsors() {
-    return (
-        <>
-            <h2 className="project-heading" style={{ paddingBottom: "20px" }}>
-                Our <strong className="purple">Sponsors</strong>
-            </h2>
-            <Row style={{ justifyContent: "center", paddingBottom: "50px" }}>
-                {sponsorsData.sponsors.map((sponsor, index) => (
-                    <Col xs={6} md={3} className="tech-icons" key={index}>
+    const sponsors = [
+        
+        {
+            href: 'https://engineering.vanderbilt.edu/ece',
+            imgSrc: 'assets/sponsors/ece.png',
+            alt: 'Electrical and Computer Engineering',
+        },
+        {
+            href: 'https://www.vanderbilt.edu/vise',
+            imgSrc: 'assets/sponsors/vise.png',
+            alt: 'Vanderbilt Institute for Surgery and Engineering',
+        },
+        {
+            href: 'https://engineering.vanderbilt.edu/cs',
+            imgSrc: 'assets/sponsors/cs.png',
+            alt: 'Computer Science',
+        },
+        {
+            href: 'https://medschool.vanderbilt.edu/brain-institute',
+            imgSrc: 'assets/sponsors/vbi.png',
+            alt: 'Vanderbilt Brain Institute',
+        },
+        {
+            href: 'https://vkc.vumc.org/vkc',
+            imgSrc: 'assets/sponsors/vkc.png',
+            alt: 'Vanderbilt Kennedy Center',
+        },
+        {
+            href: 'https://engineering.vanderbilt.edu',
+            imgSrc: 'assets/sponsors/vuse.png',
+            alt: 'Vanderbilt University School of Engineering',
+        },
+        {
+            href: 'https://www.vanderbilt.edu/datascience',
+            imgSrc: 'assets/sponsors/dsi.png',
+            alt: 'Vanderbilt Data Science Institute',
+        },
+        {
+            href: 'https://www.vumc.org/vuiis',
+            imgSrc: 'assets/sponsors/vuiis.png',
+            alt: 'Vanderbilt University Institute of Imaging Science',
+        },
+        {
+            href: 'https://engineering.vanderbilt.edu/bme',
+            imgSrc: 'assets/sponsors/bme.png',
+            alt: 'Biomedical Engineering',
+        }
+    ];
 
-                        <a
-                            href={sponsor.url}
-                            target="_blank"
-                            rel="noopener noreferrer"
-                            style={{
-                                display: "flex",
-                                justifyContent: "center",
-                                alignItems: "center",
-                                height: "100%",
-                                width: "100%"
-                            }}
-                        >
-                            <img
-                                src={sponsor.img}
-                                alt={sponsor.name}
-                                style={{
-                                    maxWidth: sponsor.dims.width,
-                                    maxHeight: sponsor.dims.height,
-                                    width: "100%",
-                                    height: "auto",
-                                    objectFit: "contain",
-                                    padding: "20px"
-                                }}
-                            />
-                        </a>
-                    </Col>
-                ))}
-            </Row >
-        </>
+    return (
+        <div className="sponsor-section">
+            <Row className="justify-content-center">
+                <Col lg={10}>
+                    <h2 className="project-pitch-heading">
+                        2024 <span className="purple">Sponsors</span>
+                    </h2>
+                    <Row style={{ justifyContent: "center", paddingBottom: "50px" }}>
+                        {sponsors.map((sponsor, index) => (
+                            <Col key={index} xs={6} md={3} lg={3} className="sponsor-icons">
+                                <a href={sponsor.href} target="_blank" rel="noopener noreferrer">
+                                    <img src={sponsor.imgSrc} alt={sponsor.alt} style={{ width: "100%", height: "auto" }} />
+                                </a>
+                            </Col>
+                        ))}
+                    </Row>
+                </Col>
+            </Row>
+        </div>
     );
 }
