+++ src/style.css
@@ -1086,3 +1086,50 @@ button:focus {
 .fork-btn-inner::after {
   display: none !important;
 }
+
+/* Sponsors Section Styles */
+
+.sponsor-section {
+  padding: 4rem 0;
+  margin-top: 2rem;
+  margin-bottom: 2rem;
+  background-image: linear-gradient(to left, rgb(27 20 41), rgb(20 15 35));
+}
+
+.sponsor-container {
+  background-color: rgba(0, 0, 0, 0.2);
+  padding: 2rem;
+  border-radius: 10px;
+  color: white;
+}
+
+.sponsor-icons {
+  font-size: 4.5em !important;
+  margin: 15px !important;
+  padding: 10px !important;
+  opacity: 0.93 !important;
+  border: 1.7px solid rgba(200, 137, 230, 0.637) !important;
+  vertical-align: middle !important;
+  text-align: center !important;
+  border-radius: 5px !important;
+  display: table !important;
+  box-shadow: 4px 5px 4px 3px rgba(89, 4, 168, 0.137) !important;
+  overflow: hidden !important;
+  transition: all 0.4s ease 0s !important;
+}
+
+@media (max-width: 767px) {
+  .sponsor-icons {
+    margin: 10px !important;
+  }
+}
+
+.sponsor-icons:hover {
+  transform: scale(1.05) !important;
+  overflow: hidden !important;
+  border: 2.2px solid rgba(197, 115, 230, 0.883) !important;
+}
+.sponsor-icon-images {
+  padding: 20px !important;
+  line-height: 1.6 !important;
+}
