--- src/styles/components/sponsor_section.css
@@ -1,49 +1,0 @@
-/* Center the entire section */
-.sponsors-section {
-    text-align: center;
-    padding: 2rem;
-  }
-  
-  /* Style for the title */
-  .sponsors-section h2 {
-    margin-bottom: 2rem;
-    font-size: 2rem;
-    font-weight: bold;
-    color: white;
-  }
-  
-  /* Grid layout for sponsors */
-  .sponsors-grid {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 2rem;
-  }
-  
-  /* Individual sponsor card */
-  .sponsor-card {
-    width: 150px;
-    text-align: center;
-  }
-  
-  /* Sponsor images */
-  .sponsor-card img {
-    width: 100%;
-    height: auto;
-    border-radius: 8px;
-    transition: transform 0.3s ease, box-shadow 0.3s ease;
-  }
-  
-  /* Hover effect on images */
-  .sponsor-card img:hover {
-    transform: scale(1.05);
-    box-shadow: 0 4px 15px rgba(0, 0, 0, 0.2);
-  }
-  
-  /* Sponsor names */
-  .sponsor-card p {
-    margin-top: 0.5rem;
-    font-size: 1rem;
-    font-weight: 500;
-  }
-
